Changed int to double in TotalExpenses classes and formatted displaying them on charts.

+++ lib/datamodels/total_expenses.dart
@@ -1,18 +1,20 @@
 import 'package:flutter/material.dart';
 
 abstract class TotalExpenses {
-  final int totalMoneyAmount;
+  final double totalMoneyAmount;
   final String name;
 
   TotalExpenses(this.totalMoneyAmount, this.name);
 }
 
 class TotalMonthlyExpenses extends TotalExpenses {
-  TotalMonthlyExpenses({@required int totalMoneyAmount, @required String name})
+  TotalMonthlyExpenses(
+      {@required double totalMoneyAmount, @required String name})
       : super(totalMoneyAmount, name);
 }
 
 class TotalCategoryExpenses extends TotalExpenses {
-  TotalCategoryExpenses({@required int totalMoneyAmount, @required String name})
+  TotalCategoryExpenses(
+      {@required double totalMoneyAmount, @required String name})
       : super(totalMoneyAmount, name);
 }
